requirements.txt
pandas
numpy
matplotlib
streamlit
openai==0.27.0
langchain==0.0.141
huggingface-hub==0.11.0
python-dotenv
